Move key filtering into a helper class
Extracted key filtering, makes the base class cleaner and makes it clear that each handler has to implement  the same Keys interface.
